Fix 18939: vpParseArgv::ARGV_CONSTANT segfault

+++ src/tools/io/vpParseArgv.cpp
@@ -173,8 +173,8 @@ vpParseArgv::parse(int *argcPtr, const char **argv, vpArgvInfo *argTable,
       infoPtr = matchPtr;
       switch (infoPtr->type) {
       case ARGV_CONSTANT:
-  *((long long *) infoPtr->dst) = 1;
-         break;
+        *((int *) infoPtr->dst) = 1;
+        break;
       case ARGV_INT:
          nargs = (uintptr_t) infoPtr->src;
          if (nargs<1) nargs=1;
